Refactor the layout of nsx-operator
