--- config/samples/securitypolicy_v1_securitypolicy.yaml
@@ -1,9 +1,0 @@
-apiVersion: nsx.vmware.com/v1alpha1
-kind: SecurityPolicy
-metadata:
-  name: securitypolicy-sample
-spec:
-  # TODO(user): Add fields here
-  foo: foo
-status:
-  status: true
